warnings.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@42548 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/BiosCommonCode/MakeAlternativeSource.cpp
@@ -907,7 +907,7 @@ static DECLCALLBACK(int) disReadOpcodeBytes(PDISCPUSTATE pDis, uint8_t offInstr,
             cbToRead = g_cbImg - offBios;
     }
     memcpy(&pDis->abInstr[offInstr], &g_pbImg[offBios], cbToRead);
-    pDis->cbCachedInstr = offInstr + cbToRead;
+    pDis->cbCachedInstr = (uint8_t)(offInstr + cbToRead);
     return VINF_SUCCESS;
 }
 
@@ -1093,7 +1093,7 @@ static RTEXITCODE DisassembleBiosImage(void)
 
     /* Final gap. */
     if (uFlatAddr < g_uBiosFlatBase + g_cbImg)
-        fRc = disCopySegmentGap(uFlatAddr, g_uBiosFlatBase + g_cbImg - uFlatAddr);
+        fRc = disCopySegmentGap(uFlatAddr, (uint32_t)(g_uBiosFlatBase + g_cbImg - uFlatAddr));
     else if (uFlatAddr > g_uBiosFlatBase + g_cbImg)
         return RTMsgErrorExit(RTEXITCODE_FAILURE, "Last segment spills beyond 1MB; uFlatAddr=%#x\n", uFlatAddr);
